==> 1 Web scraping IMDB.ipynb <==
## WebScraping página IMDB:

%%time

#Declaracion de librerias
import requests
from requests import get
from bs4 import BeautifulSoup
import pandas as pd
import numpy as np

from time import sleep
from random import randint

#Declaracion de array para almacenar resultados
xtitulo = []
xlink = []
xdescripcion = []

#Rango de películas a procesar
pag_inicio = 1
pag_fin    = 6851

#Inicio de proceso de extracción
pages = np.arange(pag_inicio, pag_fin, 50)
for page in pages:
    website1 = 'https://www.imdb.com/search/title/?title_type=feature&release_date=1979-01-01,1980-12-31&start=' + str(page) + "&ref_=adv_nxt"
    print('Procesando página: ' + str(page) + '  ' + website1)
    headers = {"Accept-Language":"es-es"}
    pagina1 = requests.get(website1, headers=headers)
    soup1 = BeautifulSoup(pagina1.text,'html.parser')  #'lxml')
    movie_div = soup1.find_all('div', class_='lister-item-content')
    
    sleep(randint(2,10))
    
    for item1 in movie_div:
        #Obtiene titulo
        title = item1.h3.a.text
        if title != None:
            xtitulo.append(title)
        else:
            xtitulo.append('Sin título')
                
        #Obtiene link
        lista = str(item1.h3.get_attribute_list) #Extrae lista completa de datos        
        
        try:
            pos = lista.index('<a href=') #busca la etiqueta href
            cadena = lista[pos+16:] #extrae toda la cadena
            pos = cadena.index('/') #encuentra la posicion del símbolo "/"
            cadena = cadena[0:pos] #extrae código de la pelicula
            cadenalink = 'https://www.imdb.com/title/' + cadena + '/?ref_=adv_li_tt' #completa cadena del link
            xlink.append(cadenalink)
        except ValueError:
            xlink.append('Sin link')
        
        #Ir al detalle de la pelicula para extraer la descripción
        website2 = cadenalink
        pagina2 = requests.get(website2, headers=headers)
        soup2 = BeautifulSoup(pagina2.text, 'html.parser')
       
        #descrip = soup2.find('span', class_='GenresAndPlot__TextContainerBreakpointXS_TO_M-sc-cum89p-0 kHlJyu')
        descrip = soup2.find('span', class_='sc-16ede01-2 gXUyNh')
        if descrip != None:
            lista = str(descrip)
            try:
                pos = lista.index('"presentation"') #busca la etiqueta presentation
                cadena = lista[pos+15:] #extrae toda la cadena
                pos = cadena.index('</span>') #encuentra la posicion del símbolo "/"
                cadena = cadena[0:pos] #extrae código de la pelicula
                xdescripcion.append(cadena)
            except ValueError:
                xdescripcion.append('Sin descripción')
        else:
            xdescripcion.append('Sin descripción')
            
            
#Mueve los arrays a un dataframe
movies = pd.DataFrame({
'titulo': xtitulo,
'link': xlink,
'descripcion': xdescripcion
})

#move dataframe a un archivo CSV
movies.to_csv('movies_Anna_1979_1980.csv', encoding='utf-8-sig')
