==> 3 Limpieza de etiquetas html.ipynb <==
### Instalación de librerias

#Instalación de libreria
#!pip install contractions

### Importar librerias

import contractions
import pandas as pd

#### Leer archivo Csv

#Archivo Input
name_file = 'movies_Ruben'
ext_file = '.csv'
name_file_input = name_file + ext_file  #Arma nombre de archivo
df_films = pd.read_csv(name_file_input) #Carga la data de las peliculas en un Dataframe

#### Declara funciones

def reemplaza_caracteres(texto):
    cadena1 = texto.replace("“", '"').replace("”", '"').replace("‘", "'").replace("’", "'").replace("'s", "s")
    cadena2 = cadena1.replace("'", "").replace('"', "")
    return cadena2

def extrae_html(texto):
    try:
        pos = texto.index('<!--') #busca la etiqueta <!--
        cadena = texto[0:pos] #extrae parte de la cadena
    except ValueError:
        cadena = texto
    return cadena

#### Procesamiento

numreg_tot = 0
numreg_condescrip = 0
numreg_sindescrip = 0
#Crea dataframe temporal
df_temporal = pd.DataFrame()
df_temporal['titulo'] = ''
df_temporal['link'] = ''
df_temporal['descripcion'] = ''
#Lee dataframe de peliculas y limpia registros
for index, row in df_films.iterrows():
    xtitulo = row['titulo']
    xlink = row['link']
    xdescripcion = row['descripcion']
    if not pd.isna(xdescripcion): #Verifica si tiene descripción
        descripcion = row['descripcion']
        posicion = 0
        try:
            posicion = descripcion.index('Synopsis') #busca texto "Synopsis" para quitarlo
        except ValueError:
            posicion = 0
            
        if posicion == 0:
            #Llama a función que reemplaza caracteres
            descrip_limpia1 = contractions.fix(descripcion)
            #Llama a función que expande contracciones a las descripciones en inglés
            descrip_limpia2 = reemplaza_caracteres(descrip_limpia1)
            #Llama a función que extrae etiquetas html
            descrip_limpia3 = extrae_html(descrip_limpia2)
            #Agrega registros al dataframe temporal para almacenar los registros que se van leyendo
            df_temporal = df_temporal.append({'titulo' : xtitulo, 'link' : xlink, 'descripcion' : descrip_limpia3} , ignore_index=True)
        numreg_condescrip = numreg_condescrip + 1
    else:
        numreg_sindescrip = numreg_sindescrip + 1
        
    numreg_tot = numreg_tot + 1 
#Exporta dataframe temporal a csv
name_file_output = name_file + '_limpio' + ext_file 
df_temporal.to_csv(name_file_output, encoding='utf-8-sig')
print('*****************************************')
print('******* Fin de proceso de limpieza ******')
print('*****************************************')
print('')
print('Número de registros con descripción: ' + str(numreg_condescrip))
print('Número de registros sin descripción: ' + str(numreg_sindescrip))
print('Número total de registros          : ' + str(numreg_tot))
