==> 5 Detección inglés.ipynb <==
## Detección de descripción en inglés

import pandas as pd
import numpy as np

movie_df = pd.read_csv('movies__Anna_traducido.csv')
movie_df.drop('Unnamed: 0', axis=1, inplace=True)
#movie_df.drop('Unnamed: 0.1', axis=1, inplace=True)
movie_df

#dejamos caer las filas que contienen valores nulos
movie_df.dropna(subset = ["traduccion"], inplace=True)
#dejamos caer las columnas que contienen los datos que no vamos a usar
movie_df.drop(["descripcion","idioma"], inplace=True, axis=1)
movie_df

%%time

# detectamos la traducción que ha quedado en inglés

from langdetect import detect

lang = []    
    
for index, row in movie_df['traduccion'].iteritems():
    movie_df['traduccion'] = movie_df['traduccion'].apply(str)
    try:
        lang = detect(row) #detecting each row
    except:
        lang = "error"
    movie_df.loc[index, 'traduccion_idioma'] = lang

movie_df

#películas con la descripción en ingles
movie_df.loc[movie_df['traduccion_idioma'] == 'en']

#películas donde no se ha podido detectar el idioma
movie_df.loc[movie_df['traduccion_idioma'] == 'error']

#dejamos caer las columnas que contienen los datos que no vamos a usar
movie_df.drop('Unnamed: 0.1', axis=1, inplace=True)
#dejamos caer las filas que contienen inglés y error
movie_df = movie_df[movie_df.traduccion_idioma != 'en']
movie_df = movie_df[movie_df.traduccion_idioma != 'error']
movie_df

# guardamos el dataset limpio
movie_df.to_csv('movies_final_Anna.csv', encoding='utf-8-sig')

### Juntamos tres datasets y analizamos los datos

import glob 
import os

# establecemos la ruta para unir varios archivos
files = os.path.join("movies_final_*.csv")
# lista de archivos combinados
files = glob.glob(files)
# unimos archivos con concat y read_csv
df = pd.concat(map(pd.read_csv, files), ignore_index=True)
# eliminamos columna Unnamed: 0 para que no aparezca la numeración individual de cada archivo
df.drop('Unnamed: 0', axis=1, inplace=True)
# guardamos el archivo conjunto
#df.to_csv('movies_final_all.csv', encoding='utf-8-sig')

# analizamos la columna descripcion_idioma
df_detect_lang = df.groupby(['traduccion_idioma']).size().sort_values(ascending=False)
df_detect_lang

# como hemos detectado más idiomas en el dataset, creamos uno nuevo solamente con descripciones en español
movies_final = df.loc[df['traduccion_idioma'] == 'es']
# eliminamos películas sin descripción
movies_final = movies_final[movies_final.traduccion != "'Sin descripción'"]
movies_final.shape

# guardamos el archivo
movies_final.to_csv('movies_final_all.csv', encoding='utf-8-sig')

movies_final.loc[movies_final['traduccion'] == "'Sin descripción'"]
