==> 7 EDA (búsqueda semántica).ipynb <==
## Analisis exploratorio de datos de NLP

import pandas as pd
import numpy as np

movie_df = pd.read_csv('movies_final_all.csv')
#dejamos caer las columnas que contienen los datos que no vamos a usar
movie_df.drop('Unnamed: 0', axis=1, inplace=True)
movie_df.drop('traduccion_idioma', axis=1, inplace=True)
movie_df

Analisis exploratorio de datos antes de limpieza (número de carácteres, número de palabras)

#dejamos caer las filas que contienen valores nulos
movie_df.dropna(subset = ["traduccion"], inplace=True)
#el número de carácteres
movie_df['trad_num_car'] = movie_df['traduccion'].str.len()
#el número de palabras
movie_df['trad_num_pal'] = movie_df['traduccion'].str.split().str.len()
#count.index = count.index.astype(str)
movie_df.sort_index(inplace=True)
movie_df

print("Número promedio de carácteres: " + str(movie_df['trad_num_car'].mean(axis = 0, skipna = True)))
print("Número promedio de palabras:   " + str(movie_df['trad_num_pal'].mean(axis = 0, skipna = True)))

## Limpieza de texto
Para poder extraer las características vectoriales del corpus es conveniente realizar primero una limpieza y pre-procedado de cada descripción traducida.\
Realizaremos el siguiente procesado:
- Introducir un espacio después de determinados signos de puntuación (".", "?") para que el tokenizado sea correcto
- Separar el texto en *tokens*
- Eliminar signos de puntuación o espacios o de longitud 1
- Lematizar el texto y pasarlo a minúsculas

import spacy
import os, re

nlp = spacy.load("es_core_news_sm")

def normaliza(texto):
    #separamos después de ciertos signos de puntuación
    texto = re.sub(r"([\.\?])", r"\1 ", texto)
    doc = nlp(texto)
    # quitamos puntuación/espacios
    tokens = [t for t in doc if not t.is_punct and not t.is_space and len(t.text)>1]
    palabras = []
    for t in tokens:
        palabras.append(t.lemma_.lower())
    # juntamos de nuevo en una cadena
    salida = ' '.join(palabras)
    
    return salida

%%time

#normalizamos la descripción
movie_df['traduccion_limpia'] = movie_df['traduccion'].apply(normaliza)
movie_df

# guardamos el archivo
movie_df.to_csv('movies_bus_semantica.csv', encoding='utf-8-sig')

#sample = movie_df[:15]
#sample

#%%time

#normalizamos la descripción
#sample['traduccion_limpia'] = sample['traduccion'].apply(normaliza)
#sample

Analisis exploratorio de datos después de limpieza (número de carácteres, número de palabras)

#el número de carácteres
movie_df['trad_limpia_num_car'] = movie_df['traduccion_limpia'].str.len()
#el número de palabras
movie_df['trad_limpia_num_pal'] = movie_df['traduccion_limpia'].str.split().str.len()
#count.index = count.index.astype(str)
movie_df.sort_index(inplace=True)

movie_df

print("Número promedio de carácteres: " + str(movie_df['trad_limpia_num_car'].mean(axis = 0, skipna = True)))
print("Número promedio de palabras:   " + str(movie_df['trad_limpia_num_pal'].mean(axis = 0, skipna = True)))
